Added new Authentication class, with support for User Account, login/logout plus various Events.  Currently unable to fully verify it working, as a title needs to be published on the Epic Store for this to work.
Updated the Manager to provide access to the Authentication instance.  This needs to be further updated to support multiple account logins, as the SDK allows for this.
Updated the Player Controller (Blueprint) to handle Login and Logout actions, plus various login event Bindings.

Updated to use VS2017.
(Can easily be reverted by changing the Editor Preferences Source Code Editor, and the Project Settings Compiler version)